+++ buttons/sports_button.py
@@ -1,0 +1,9 @@
+from buttons.toggle_button import ToggleButton
+from shared import user_preferences
+
+
+class SportsButton(ToggleButton):
+    def __init__(self, key, text, action):
+        super().__init__(text, action)
+        self.setChecked(user_preferences.sports_enabled[key])
+        self.setFixedSize(350, 120)
